% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br.R
\name{br}
\alias{br}
\title{Browse data}
\usage{
br(x, title)
}
\arguments{
\item{x}{An \code{R} object coercible into a data frame.}

\item{title}{Optional title for viewer window.}
}
\description{
\code{br()} is an alias for \code{utils::View()} and invokes the data viewer. See \code{utils::\link[View]{View()}} for details.
}
