% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gin.R
\name{\%gin\%}
\alias{\%gin\%}
\title{A reimagination of \code{\%in\%} for partial string matching}
\usage{
pattern \%gin\% x
}
\arguments{
\item{pattern}{Character string to be matched.}

\item{x}{\code{R} object to be matched against. The object must be a character vector or an object coercible by \code{as.character()} to a character vector.}
}
\description{
\code{\%gin\%} is a reimagination of \code{\%in\%} using \code{grepl()} for partial string matching.
}
\examples{
# \%in\% evaluates to FALSE because it looks for full string matches
"t" \%in\% "tonic"

# \%gin\% evaluates to TRUE
"t" \%gin\% "tonic"

}
\references{
\code{\%gin\%} was first written for \href{https://github.com/ivelasq}{@ivelasq}'s \href{https://github.com/ivelasq/r-data-recipes}{r-data-recipes} GitHub repository.
}
