% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{ta}
\alias{ftab}
\alias{tab1}
\alias{tab2}
\title{Create tables of frequencies}
\usage{
tab(x, ..., m = TRUE)

ta(x, ..., m = TRUE)

ftab(x, ..., m = TRUE)

tab1(x, ..., m = TRUE)

tab2(x, ..., m = TRUE)
}
\arguments{
\item{x}{A vector, data.frame, or tibble.}

\item{...}{A comma separated list of unquoted variable names.}

\item{m}{If \code{TRUE} (the default), the tables will report missing values.}
}
\value{
A tibble containing a table of frequencies for the variables listed in \code{...}
}
\description{
\code{tab()} creates n-way tables of frequencies in the \code{R} console, similar to those created by Stata's \code{tabulate} function. When three or more variables are passed to \code{tab()}, only flat tables are displayed. \code{ta()} is a shortened alias for \code{tab()}.

\code{ftab()} creates only flat tables of frequencies.

The convenience functions \code{tab1()} and \code{tab2()} are inspired by functions of the same name in Stata. They allow rapid tabulation of a set of variables. \code{tab1()} creates one-way tables of frequencies for each listed variable. \code{tab2()} creates two-way tables of frequencies for all listed variable combinations.
}
\details{
If a single variable is passed to \code{tab()}, a table of frequencies is printed (with a total row and columns 'Freq.', 'Percent', and 'Cum.').

If two variables are passed to \code{tab()}, a special 2x2 contingency table is printed (with a total row and a total column).

If three or more variables are passed to \code{tab()}, a flat contingency table is printed (with columns 'Freq.', 'Percent', and 'Cum.').

Invisibly returned tibbles exclude total rows and columns to avoid collision of variable classes.
}
\examples{
# one-way table of frequencies
mtcars \%>\% tab(cyl)

# two-way table of frequencies (a special 2x2 contingency table)
mtcars \%>\% tab(cyl, gear)

# flat contingency tables of three (or more) variables
mtcars \%>\% tab(cyl, gear, am)

# tables wider than the R console are automatically wrapped
mtcars \%>\% tab(cyl, gear, am, vs)

# ftab() displays only flat contingency tables (here, with two variables)
mtcars \%>\% ftab(cyl, gear)

# tab1() displays one-way tables for each variable
mtcars \%>\% tab1(cyl, gear)

# tab2() displays two-way tables for all variable combinations
mtcars \%>\% tab2(cyl, gear, am)

# ta() is a shortened alias for tab(), inspired by Stata
mtcars \%>\% ta(gear)

}
\seealso{
The statar package by Matthieu Gomez provides a \code{tab()} function with output similar to tabr's \code{ftab()}. Both packages use a variant of \href{https://github.com/gvelasq-r/statascii}{statascii()} to format tables for display in the \code{R} console. Differences between the packages include:
\itemize{
\item tabr displays tables in tidyverse colors: grey for block drawing characters and red for \code{NA}s.
\item tabr allows for tabulation of named and unnamed vectors.
\item tabr implements automatic table wrapping for tables wider than the \code{R} console.
\item tabr's \code{tab()} and \code{ftab()} display a total row with total frequencies for one-way tabulations.
\item tabr's \code{tab()} displays a special 2x2 contingency table for two-way tabulations (flat two-way tables are available with \code{ftab()}).
\item tabr's convenience functions \code{tab1()} and \code{tab2()} allow for rapid tabulation of a set of variables into either one- or two-way tables.
}

The janitor package by Sam Firke provides the \code{tabyl()} function for SPSS-like tables of frequencies and adornments.

Base \code{R} provides the \code{base::ftable()} and \code{stats::xtabs()} functions for unadorned tables of frequencies.
}
