% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabr-package.R
\docType{package}
\name{tabr-package}
\alias{tabr}
\alias{tabr-package}
\title{tabr: Create Tables of Frequencies}
\description{
Functions to create tables of frequencies.
}
\author{
\strong{Maintainer}: Gustavo Velásquez \email{gvelasq2github@gmail.com}

}
\keyword{internal}
