% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{ta}
\alias{ftab}
\alias{tab1}
\alias{tab2}
\title{Create tables of frequencies}
\usage{
tab(x, ..., m = TRUE)

ta(x, ..., m = TRUE)

ftab(x, ..., m = TRUE)

tab1(x, ..., m = TRUE)

tab2(x, ..., m = TRUE)
}
\arguments{
\item{x}{A vector, data.frame, or tibble.}

\item{...}{Comma separated list of unquoted variable names.}

\item{m}{If \code{TRUE} the tables will report missing values.}
}
\value{
A tibble containing a table of frequencies for the variables listed in \code{...}
}
\description{
\code{tab()} and \code{ta()} are synonyms and create one-way and two-way tables of frequencies in the \code{R} console, similar to those created by Stata's \code{tabulate} function. When three or more variables are provided, these functions print flat tables of frequencies.

\code{ftab()} creates only flat tables of frequencies.

The convenience functions \code{tab1()} and \code{tab2()} are inspired by functions of the same name in Stata. They allow rapid tabulation of a set of variables during interactive data analysis. \code{tab1()} creates one-way tables of frequencies for each listed variable. \code{tab2()} creates two-way tables of frequencies for all listed variable combinations.
}
\details{
If a single variable is passed to \code{tab()}, a table of frequencies is printed (with a total row and columns 'Freq.', 'Percent', and 'Cum.').

If two variables are passed to \code{tab()}, a 2x2 contingency table is printed (with a total row and a total column).

If three or more variables are passed to \code{tab()}, a flat contingency table is printed (with columns 'Freq.', 'Percent', and 'Cum.').

Invisibly returned tibbles exclude total rows and columns to avoid collision of variable classes.
}
\examples{
# setup
library(dplyr)

# one-way table of frequencies
mtcars \%>\% tab(cyl)

# two-way table of frequencies (a 2x2 contingency table)
mtcars \%>\% tab(cyl, gear)

# flat contingency tables of three (or more) variables
mtcars \%>\% tab(cyl, gear, am)

# tables wider than the R console are automatically wrapped
mtcars \%>\% tab(cyl, gear, am, vs)

# ftab() displays only flat contingency tables (here, with two variables)
mtcars \%>\% ftab(cyl, gear)

# tab1() displays one-way tables for each variable
mtcars \%>\% tab1(cyl, gear)

# tab2() displays two-way tables for all variable combinations
mtcars \%>\% tab2(cyl, gear, am)

# ta() is a shortened alias for tab(), inspired by Stata
mtcars \%>\% ta(gear)

}
\seealso{
\code{statar::tab()} by Matthieu Gomez is similar to \code{tabr::ftab()} yet is implemented differently and provides different arguments. Like \code{tabr}, \code{statar} uses a variant of \href{https://github.com/gvelasq-r/statascii}{statascii()} to format tables for display in the \code{R} console.

\code{janitor::tabyl()} creates SPSS-like tabulations and adornments that are displayed using \code{tibble::tibble()} in the \code{R} console.

\code{base::ftable()}, \code{stats::xtabs()} are base \code{R} solutions for creating tables of frequencies.
}
