% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabr-package.R
\docType{package}
\name{tidytab-package}
\alias{tidytab}
\alias{tidytab-package}
\title{tidytab: Create Tables of Frequencies}
\description{
Functions to create tables of frequencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gvelasq/tidytab}
  \item Report bugs at \url{https://github.com/gvelasq/tidytab/issues}
}

}
\author{
\strong{Maintainer}: Gustavo Velásquez \email{gvelasq2github@gmail.com}

}
\keyword{internal}
